Added Message and Purchase Controllers, finished User and Game controllers, removed unused delete method from PurchaseRepo, Modified Message and Purchase Service for return codes.

+++ backend/src/main/java/com/github/rahultole06/TheGamesClub/backend/controllers/UserController.java
@@ -23,6 +23,7 @@ public class UserController {
   }
 
   public int createUser(String username, String password) {
+    if (username == null || password == null) { return -1; }
     try {
       User newUser = new User(username, password);
       return userService.createUser(newUser);
@@ -32,6 +33,7 @@ public class UserController {
   }
 
   public int updateUser(String username, String oldPassword, String newPassword) {
+    if (username == null || oldPassword == null || newPassword == null) { return -1; }
     try {
       User newUser = new User(username, newPassword);
       return userService.updateUser(newUser, oldPassword);
@@ -41,6 +43,7 @@ public class UserController {
   }
 
   public int deleteUser(String username, String password) {
+    if (username == null || password == null) { return -1; }
     try {
       return userService.deleteUser(username, password);
     } catch (EntityNotFoundException | IllegalArgumentException e) {
@@ -49,6 +52,7 @@ public class UserController {
   }
 
   public User getUser(String username) {
+    if (username == null) { return null; }
     try {
       return userService.getUserByUsername(username);
     } catch (EntityNotFoundException e) {
@@ -57,6 +61,7 @@ public class UserController {
   }
 
   public List<Purchase> getUserPurchases(String username) {
+    if (username == null) { return null; }
     try {
       return userService.getUserPurchases(username);
     } catch (EntityNotFoundException e) {
@@ -65,6 +70,7 @@ public class UserController {
   }
 
   public int buyGame(String username, Purchase purchase) {
+    if (username == null || purchase == null) { return -1; }
     try {
       return userService.buyGame(username, purchase);
     } catch (EntityNotFoundException e) {
@@ -73,6 +79,7 @@ public class UserController {
   }
 
   public List<Game> getUserGamesAuthored(String username) {
+    if (username == null) { return null; }
     try {
       return userService.getUserGamesAuthored(username);
     } catch (EntityNotFoundException e) {
@@ -81,6 +88,7 @@ public class UserController {
   }
 
   public int addUserGamesAuthored(String username, Game game) {
+    if (username == null || game == null) { return -1; }
     try {
       return userService.addUserGamesAuthored(username, game);
     } catch (EntityNotFoundException e) {
@@ -89,6 +97,7 @@ public class UserController {
   }
 
   public List<Message> getUserMessages(String username) {
+    if (username == null) { return null; }
     try {
       return userService.getUserMessages(username);
     } catch (EntityNotFoundException e) {
@@ -97,6 +106,7 @@ public class UserController {
   }
 
   public int addUserMessages(String username, Message message) {
+    if (username == null || message == null) { return -1; }
     try {
       return userService.addUserMessages(username, message);
     } catch (EntityNotFoundException e) {
